backend/requirements.txt
python-dotenv
uvicorn
fastapi
google-generativeai 
python-multipart